# docs/api-contract.yaml
openapi: 3.1.0
info:
  title: KSA Attendance System API
  version: 0.2.0
  summary: September 2026 implementation baseline between the production frontend and backend.
  description: |
    This contract captures approved API capabilities and stable domain outcomes for
    implementation. Endpoint names may change through a documented superseding decision, but the
    authorization, state, privacy, idempotency, and source-of-truth guarantees must remain.

    The permanent QR points to a stable frontend route such as `/check-in`; it is not an
    API credential. The frontend obtains an explicit browser location and submits it to
    the backend, which makes the final attendance decision using server time.

    The production browser uses the same public origin for the UI and `/api/*`. Cookie-authenticated
    state changes also require the implementation's CSRF/origin control. Abuse-sensitive operations
    may return the documented `RATE_LIMITED` response and safe `Retry-After` under DEC-036.
servers:
  - url: /api/v1
    description: Same-origin API
...
security:
  - sessionCookie: []
paths:
...
  /admin/roster:
    get:
      tags: [Roster]
      summary: List and filter authorized roster entries
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RosterEntryStatus"
        - name: query
          in: query
          schema:
            type: string
            maxLength: 100
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Admin-only roster view
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RosterEntry"
                  nextCursor:
                    type: [string, "null"]
        "403":
          $ref: "#/components/responses/AuthorizationError"
    post:
      tags: [Roster]
      summary: Validate and import an approved participant roster CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, reason]
              properties:
                file:
                  type: string
                  format: binary
                reason:
                  type: string
                  minLength: 3
                  maxLength: 1000
      responses:
        "201":
          description: Import committed after whole-file validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterImportResultEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /admin/roster/{rosterEntryId}:
    patch:
      tags: [Roster]
      summary: Correct, disable, or restore a roster entry without deleting history
      parameters:
        - $ref: "#/components/parameters/RosterEntryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RosterEntryUpdateRequest"
      responses:
        "200":
          description: Updated and audited roster entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterEntryEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /admin/photo-change-requests:
    get:
      tags: [Identification photos]
      summary: List identification-photo replacement requests for Admin review
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReviewStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Admin-only photo-change queue
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/PhotoChangeRequest"
        "403":
          $ref: "#/components/responses/AuthorizationError"
  /admin/photo-change-requests/{requestId}:
    get:
      tags: [Identification photos]
      summary: Get current and candidate photo access for an Admin review
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Admin-only detail with few-minute presigned URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoChangeRequestEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /admin/photo-change-requests/{requestId}/approve:
    post:
      tags: [Identification photos]
      summary: Admin-only activation of the candidate identification photo
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
      responses:
        "200":
          description: Approved or already-final request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoChangeRequestEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /admin/photo-change-requests/{requestId}/reject:
    post:
      tags: [Identification photos]
      summary: Admin-only rejection of a candidate identification photo
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
      responses:
        "200":
          description: Rejected or already-final request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoChangeRequestEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /admin/roles/course-representative:
    put:
      tags: [Administration]
      summary: Replace the single active Course Representative
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantId, reason, currentPassword]
              properties:
                participantId:
                  type: string
                  format: uuid
                reason:
                  type: string
                  minLength: 3
                  maxLength: 1000
                currentPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        "200":
          description: Updated role assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /admin/roles/admins:
    get:
      tags: [Administration]
      summary: List active Administrators
      responses:
        "200":
          description: Active Administrators
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/AuthorizationError"
    post:
      tags: [Administration]
      summary: Grant Administrator role after confirmation/recent re-authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, reason, currentPassword]
              properties:
                userId:
                  type: string
                  format: uuid
                reason:
                  type: string
                  minLength: 3
                  maxLength: 1000
                currentPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        "201":
          description: Administrator role granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /admin/roles/admins/{userId}/revoke:
    post:
      tags: [Administration]
      summary: Revoke Administrator role without allowing zero active Administrators
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, currentPassword]
              properties:
                reason:
                  type: string
                  minLength: 3
                  maxLength: 1000
                currentPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        "204":
          description: Administrator role revoked
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ksa_session
      description: Secure, HttpOnly, host-only, SameSite=Lax server session; raw value is never stored. Idle expiry is 30 days and absolute expiry is 90 days.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RosterEntryId:
      name: rosterEntryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AuthorizationError:
      description: Current actor is not allowed to perform this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFoundError:
      description: Resource is absent or intentionally undisclosed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ConflictError:
      description: Current domain state conflicts with the requested command
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Role:
      type: string
      enum: [PARTICIPANT, COURSE_REP, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    User:
      type: object
      required:
        [id, fullName, email, emailVerifiedAt, roles, status, registeredAt]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        phone:
          type: [string, "null"]
          description: Normalized and unique for participant accounts
        email:
          type: string
          format: email
        serialNumber:
          type: [string, "null"]
          pattern: "^KSA-[0-9]{2}$"
        roles:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Role"
        status:
          $ref: "#/components/schemas/UserStatus"
        emailVerifiedAt:
          type: [string, "null"]
          format: date-time
        registeredAt:
          type: string
          format: date-time
    RosterEntryStatus:
      type: string
      enum: [UNCLAIMED, CLAIMED, DISABLED]
    RosterEntry:
      type: object
      required:
        [
          id,
          serialNumber,
          fullName,
          enrollmentEffectiveDate,
          status,
          version,
          createdAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        serialNumber:
          type: string
          pattern: "^KSA-[0-9]{2}$"
        fullName:
          type: string
        email:
          type: [string, "null"]
          format: email
        phone:
          type: [string, "null"]
        enrollmentEffectiveDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/RosterEntryStatus"
        claimedByUserId:
          type: [string, "null"]
          format: uuid
        claimedAt:
          type: [string, "null"]
          format: date-time
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
    RosterEntryUpdateRequest:
      type: object
      required: [expectedVersion, reason]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        fullName:
          type: string
          minLength: 2
          maxLength: 150
        email:
          type: [string, "null"]
          format: email
        phone:
          type: [string, "null"]
        enrollmentEffectiveDate:
          type: string
          format: date
        status:
          type: string
          enum: [UNCLAIMED, CLAIMED, DISABLED]
          description: Restore must preserve prior claim ownership; a claimed entry cannot be silently unclaimed and disputes use an explicit Admin process.
        reason:
          type: string
          minLength: 3
          maxLength: 1000
      additionalProperties: false
    RosterImportResult:
      type: object
      required: [importId, createdCount, updatedCount, skippedCount]
      properties:
        importId:
          type: string
          format: uuid
        createdCount:
          type: integer
          minimum: 0
        updatedCount:
          type: integer
          minimum: 0
        skippedCount:
          type: integer
          minimum: 0
        warnings:
          type: array
          items:
            type: string
    ParticipantSummary:
      type: object
      required: [id, fullName, serialNumber, status, roles]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        serialNumber:
          type: string
          pattern: "^KSA-[0-9]{2}$"
        status:
          $ref: "#/components/schemas/UserStatus"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
...
    ReviewStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, EXPIRED, SUPERSEDED]
...
    ReviewDecisionRequest:
      type: object
      required: [expectedVersion, reason]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        reason:
          type: string
          minLength: 3
          maxLength: 1000
...
    PhotoChangeRequest:
      type: object
      required: [id, participant, status, requestedAt, version]
      properties:
        id:
          type: string
          format: uuid
        participant:
          $ref: "#/components/schemas/ParticipantSummary"
        status:
          $ref: "#/components/schemas/ReviewStatus"
        note:
          type: [string, "null"]
        requestedAt:
          type: string
          format: date-time
        reviewedAt:
          type: [string, "null"]
          format: date-time
        reviewedBy:
          type: [string, "null"]
          format: uuid
        decisionReason:
          type: [string, "null"]
        currentPhotoAccessUrl:
          type: [string, "null"]
          description: Admin detail only; presigned for a few minutes
        candidatePhotoAccessUrl:
          type: [string, "null"]
          description: Admin detail only; presigned for a few minutes
        version:
          type: integer
          minimum: 1
...
    Error:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - STALE_STATE
            - DUPLICATE_EMAIL
            - DUPLICATE_PHONE
            - DUPLICATE_SERIAL
            - ROSTER_ENTRY_NOT_ELIGIBLE
            - ROSTER_CLAIM_CONFLICT
            - LAST_ADMIN_REQUIRED
            - RECENT_REAUTHENTICATION_REQUIRED
            - SELF_APPROVAL_FORBIDDEN
            - SESSION_ALREADY_OPEN
            - SESSION_NOT_OPEN
            - SESSION_CLOSED
            - SESSION_EXTENSION_LIMIT
            - DEVICE_CHANGE_REQUIRED
            - LOCATION_UNCERTAIN
            - LOCATION_PERMISSION_DENIED
            - LOCATION_TIMEOUT
            - LOCATION_UNAVAILABLE
            - LOCATION_UNSUPPORTED
            - CLEARLY_REMOTE
            - MANUAL_VERIFICATION_NOT_ALLOWED
            - MANUAL_VERIFICATION_EXPIRED
            - PHOTO_TOO_LARGE
            - PHOTO_UNSUPPORTED
            - RATE_LIMITED
            - SERVICE_TEMPORARILY_UNAVAILABLE
        message:
          type: string
        correlationId:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    UserEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/User"
...
    PhotoChangeRequestEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/PhotoChangeRequest"
    RosterEntryEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/RosterEntry"
    RosterImportResultEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/RosterImportResult"
...